refactor(transport/session): split dispatch_message benchmarks into own bench file
Move SessionManager::dispatch_message benchmarks and their supporting
infrastructure (FlumeSink, make_manager_with_sessions) from session_bench.rs
into a dedicated dispatch_message_bench.rs, leaving session_bench.rs with only
the pre-existing HoprSession wire-format benchmarks.
